% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_circle.R
\name{buffer_circle}
\alias{buffer_circle}
\title{A circule function}
\usage{
buffer_circle(point, radius, ...)
}
\arguments{
\item{point}{Centre point of the buffer}

\item{radius}{numeric}
}
\description{
Creates a circular buffer. Wrapper around sf::st_buffer()
}
\examples{
buffer_circle()
}
\keyword{cats}
