% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_wedge.R
\name{buffer_wedge}
\alias{buffer_wedge}
\title{A wedge Function}
\usage{
buffer_wedge(point, radius, degree, degree_width)
}
\arguments{
\item{point}{An sf variable, \code{\link[sf]{st_is(point, "POINT")}} must be TRUE.}

\item{radius}{numeric}

\item{degree}{numeric; must be between 0 and 360. The angle at which the wedge is centred (clockwise).}

\item{degree_width}{numeric; must be between 0 and 360. The width of the wedge.}
}
\description{
Creates a wedge polygon
}
\examples{
buffer_wedge()
}
\keyword{wedge}
