% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_windrose.R
\name{buffer_windrose}
\alias{buffer_windrose}
\title{A wedge Function}
\usage{
buffer_windrose(
  point,
  wind_frequency_df,
  radius = 100,
  m_s_scaling_normalization = 1,
  width_factor = 2
)
}
\arguments{
\item{radius}{Do you love cats? Defaults to TRUE.}

\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
wedge()
}
\keyword{cats}
