% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_sample}
\alias{wind_sample}
\title{Wind Data of Hong Kong's King's Park meteorological station for 2010 in hourly resolution}
\format{Dataframe with 8549 rows and 3 variables:
\describe{
\item{Time}{datetime}
\item{ws}{wind speed in m/s}
\item{wd}{dominating wind directions}
}}
\source{
\url{https://www.hko.gov.hk/en/wxinfo/aws/kpinfo.htm}
}
\usage{
wind_sample
}
\description{
Wind Data of Hong Kong's King's Park meteorological station for 2010 in hourly resolution
}
\keyword{datasets}
